Separate spline_filter and zoom_shift tests
